Unconditionally increase flash buffer size, enough to handle 4096-bit RSA imports.
This is rather crude, and not quite minimal. It also appears that it
will require this much space at some point no matter key import size
(when built with this patch), not just for 4096-bit imported keys.

But I am not familiar enough with the platform to do this better.